dynamic random testing

+++ test/CallingApp_server_test.py
@@ -13,6 +13,7 @@ import requests
 import random
 import shutil
 import linecache #for searching for sampled lines in combinatoric file
+from random import randint
 
 from os import environ
 try:
@@ -93,7 +94,7 @@ class CallingAppTest(unittest.TestCase):
 
 	#To run tests faster (for Mika) use bool
 	#Real tests should have bool=FALSE
-	isMika = False 
+	isMika = True 
 
 	if (not isMika):
 		#Set up test narrative
@@ -149,8 +150,8 @@ class CallingAppTest(unittest.TestCase):
 		#suffix = int(time.time() + 1000)
 		#wsName = "test_media_Calling_" + str(suffix)
 		#ret = self.wsClient.create_workspace({'workspace':wsName})
-		wsName = 'mikaelacashman:narrative_1498142336734'#ExploreFBAParams (production)
-		
+		wsName = 'mikaelacashman:narrative_1500516864725'#EColi_test (production)	
+	
 		#Set up imported modules
 		fba = fba_tools(os.environ['SDK_CALLBACK_URL'])
 		genomeUtil = GenomeFileUtil(os.environ['SDK_CALLBACK_URL'])
@@ -169,15 +170,25 @@ class CallingAppTest(unittest.TestCase):
 	status_file = open(status_filename,"w+")
 	ID_filename = self.cfg['scratch'] + "/ID.out"
 	ID_file = open(ID_filename,"w+")	
-	sample_filename = self.cfg['scratch'] + "/samples.out"
-	sample_file = open(sample_filename,"w+")	
 
 	#=================================================================
-	#============== Change Test File Here ============================
+	#============== Change Range of Values here ======================
 	#=================================================================
-	allID_filename = "IDs_overnight_test1.out"
-	allID_path = os.path.join(self.cfg['scratch'],allID_filename)
-	shutil.copy(os.path.join("test_files",allID_filename),allID_path)
+	FVA_range =      [1,0]
+	minFlux_range =  [0,1]
+	singleKO_range = [0,1]
+	expthresh_range = [0,.5,1]
+	expuncer_range =  [0,.5,1]
+	actcoeff_range =  [0,.5,1]
+	maxC_range = [0,1,5,10,20,30,40,50,100,1000]
+	maxN_range = [0,1,5,10,20,30,40,50,100,1000]
+	maxP_range = [0,1,5,10,20,30,40,50,100,1000]
+	maxS_range = [0,1,5,10,20,30,40,50,100,1000]
+	maxO_range = [0,1,5,10,20,30,40,50,100,1000]
+	#hidden params
+	thermo_range = [0,1]
+	minmed_range = [0,1]
+	allrev_range = [0,1]
 
 	#Set up basic fbaparams
 	fbaparams = {
@@ -186,90 +197,73 @@ class CallingAppTest(unittest.TestCase):
         	"media_id":  "CDG.media",
         	"target_reaction":  "bio1",
         	"fba_output_id":  "test_file_params",
-        	"fva":  1, 
-        	"minimize_flux":  1,  
-        	"simulate_ko":  1,  
-        	"feature_ko_list":  [], 
-        	"reaction_ko_list":  "", 
-        	"custom_bound_list":  [], 
-        	"media_supplement_list":  "", 
-        	"exp_threshold_percentile":  0.5,
-        	"exp_threshold_margin":  0.1,
-        	"activation_coefficient":  0.5 
 	}
 	#fbaparams['custom_bound_list']=["-100;cpd00009;1","-100;cpd00029;1","-100;cpd00084;10","-100;cpd00011;10"]
-
-	#Random Sample from file
-	#set and print a random seed
-	#generate the list to sample
-	#sample = random.sample(xrange(0,2),NUM_TESTS)
-	#sample_file.write(str(sample))
-	#sample_file.close()
-	
-	#Run random tests from exisiting file
-	rand_filename = "IDs_172mil_V1.out"
-	rand_filepath = os.path.join(self.cfg['scratch'],rand_filename)
-	shutil.copy(os.path.join("test_files",rand_filename),rand_filepath)
-	#rand_filenamecopied = self.cfg['scratch'] + "./IDs_172milTest.out"
-	rand_file = open(rand_filepath,"r")
-	#rand_file = open(rand_filenamecopied,"r")
-	#read in lines to a data structure
-	tests = rand_file.readlines()
-	print("Number of tests to run: " + str(len(tests)))	
 
 	count = 0
 	returnVal = {}
 	IDs = []
 	OVs = []
 	times = []
-
 		
-	NUM_TESTS = 20
+	#=================================================================
+	#============== Set number of runs here ==========================
+	#=================================================================
+	NUM_TESTS = 4
+
 	print("Begining loop.  Elapsed time: " + str(time.time()-FullStart))
-	#for samp in sample:
 	for t in range(0,NUM_TESTS):
 		print("loop: " + str(count) + "	   elapsed time: " + str(time.time()-FullStart))
-		#ID=linecache.getline(allID_path,samp+1).strip()
-		#params=ID.split(",")
-		params = tests[t].strip().split(",")
-		IDs.append(tests[t])	
-		fbaparams['fva']	  =float(params[0])
-		fbaparams['minimize_flux']=float(params[1])
-		fbaparams['simulate_ko']  =float(params[2])
-		fbaparams['exp_threshold_percentile']=float(params[3])
-		fbaparams['exp_threshold_margin']    =float(params[4])
-		fbaparams['activation_coefficient']  =float(params[5])
-		fbaparams['max_c_uptake']=float(params[6])
-		fbaparams['max_n_uptake']=float(params[7])
-		fbaparams['max_p_uptake']=float(params[8])
-		fbaparams['max_s_uptake']=float(params[9])
-		fbaparams['max_o_uptake']=float(params[10])
-		fbaparams['thermodynamic_constraints']=float(params[11])
-		fbaparams['find_min_media']=float(params[12])
-		fbaparams['all_reversible']=float(params[13])	
+		#randomly set the params
+		fbaparams['fva']	  =FVA_range[randint(0,1)]
+		fbaparams['minimize_flux']=minFlux_range[randint(0,1)]
+		fbaparams['simulate_ko']  =singleKO_range[randint(0,1)]
+		fbaparams['exp_threshold_percentile']=expthresh_range[randint(0,2)]
+		fbaparams['exp_threshold_margin']    =expuncer_range[randint(0,2)]
+		fbaparams['activation_coefficient']  =actcoeff_range[randint(0,2)]
+		fbaparams['max_c_uptake']=maxC_range[randint(0,9)]
+		fbaparams['max_n_uptake']=maxN_range[randint(0,9)]
+		fbaparams['max_p_uptake']=maxP_range[randint(0,9)]
+		fbaparams['max_s_uptake']=maxS_range[randint(0,9)]
+		fbaparams['max_o_uptake']=maxO_range[randint(0,9)]
+		fbaparams['thermodynamic_constraints']=thermo_range[randint(0,1)]
+		fbaparams['find_min_media']=minmed_range[randint(0,1)]
+		fbaparams['all_reversible']=allrev_range[randint(0,1)]
 
-		#ID_file.write(",".join(params))
-		#ID_file.write("\n")
-									
 		##Make the call to FBA
 		start = time.time()
 		##Call to FBA
 		files = fba.run_flux_balance_analysis(fbaparams)
 		new_fba_ref = files['new_fba_ref']
-		returnVal['OV'] = str(self.wsClient.get_object(
+		#save time, OV, and params
+		IDs.append(
+			str(fbaparams['fva']) + "," + 
+			str(fbaparams['minimize_flux']) + "," +
+			str(fbaparams['simulate_ko']) + "," + 
+			str(fbaparams['exp_threshold_percentile']) + "," +
+			str(fbaparams['exp_threshold_margin']) + "," + 
+			str(fbaparams['activation_coefficient']) + "," +
+			str(fbaparams['max_c_uptake']) + "," +
+			str(fbaparams['max_n_uptake']) + "," +
+			str(fbaparams['max_p_uptake']) + "," +
+			str(fbaparams['max_s_uptake']) + "," +
+			str(fbaparams['max_o_uptake']) + "," +
+			str(fbaparams['thermodynamic_constraints']) + "," +
+			str(fbaparams['find_min_media']) + "," +
+			str(fbaparams['all_reversible'])
+		)
+		OVs.append(str(self.wsClient.get_object(
 			{'workspace':wsName,
-			'id':fbaparams['fba_output_id']})['data']['objectiveValue'])
-		returnVal['time'] = str(time.time()-start)       	
+			'id':fbaparams['fba_output_id']})['data']['objectiveValue']))
+		times.append(str(time.time()-start))       	
 		count+=1
-		#OV_file.write("%s\n" %returnVal['OV'])
-		#status_file.write("%s\n" %returnVal['time'])
-		OVs.append(returnVal['OV'])
-		times.append(returnVal['time'])		
-
+		
+		print(OVs)
 		#print IDs every 10 runs
-		if count%10 == 0 and count>1:
-			for j in range(0,9):
-				ID_file.write(IDs[j])
+		if count%2 == 0 and count>1:
+			print("On the 10th - writing")
+			for j in range(0,1):
+				ID_file.write("%s\n" %IDs[j])
 				OV_file.write("%s\n" %OVs[j])
 				status_file.write("%s\n" %times[j])
 			IDs = []
@@ -288,9 +282,10 @@ class CallingAppTest(unittest.TestCase):
 	OV_file.close()
 	status_file.close()
 	ID_file.close()
-	#Delete test workspace
-	self.wsClient.delete_workspace({'workspace': wsName})
-        print('Test workspace was deleted')
+	if (not isMika):
+		#Delete test workspace
+		self.wsClient.delete_workspace({'workspace': wsName})
+        	print('Test workspace was deleted')
 	print('Total elapsed time: ' + str(time.time()-FullStart))
 	#DONE
